Abstracted methods from DSETestGenerator and moved to utils: - Solver call - BuildQuery
